+++ invert.cc
@@ -2,12 +2,12 @@
 
 #include <stdio.h>
 
-float inverter::x( float f1 )
+double inverter::x( double f1 )
 {
-  float f2, x1 = maxx / 2.0;
-  float w = maxx / 4.0;
+  double f2, x1 = maxx / 2.0;
+  double w = maxx / 4.0;
 
-  // no point in doing more than 32 iterationd for float (precission)
+  // no point in doing more than 32 iterationd for double (precission)
   for( int i = 0; i < 32; i++ )
   {
     f2 = f(x1) / maxf;
@@ -22,7 +22,7 @@ float inverter::x( float f1 )
   return x1;
 }
 
-float massInverter::f( float x )
+double massInverter::f( double x )
 {
   return (   100.088 
            + 0.112798 * erff(-5.56257 + 0.0471405 * x) 
@@ -33,8 +33,8 @@ float massInverter::f( float x )
          ) / 200.1756;
 }
 
-float energyInverter::f( float x )
+double energyInverter::f( double x )
 {
-  float x1 = x / ( 1.0 - A / 234.0 );
+  double x1 = x / ( 1.0 - A / 234.0 );
   return ( -0.00014122 + (0.00014122 -7.12299E-7 * x1 ) * exp(0.0886603 * x1) ) / 127.216;
 }
