+++ sample_dynamic.h
@@ -11,8 +11,8 @@ struct sampleDynamic : sampleLayers {
 
   virtual void averages( const ionBase *_pka );
 
-  sampleDynamic( float x, float y, float z ): sampleLayers( x, y, z) { bc[0] = CUT; bc[1] = PBC; bc[2] = PBC; };
-  virtual materialBase* lookupMaterial( float* pos );
+  sampleDynamic( double x, double y, double z ): sampleLayers( x, y, z) { bc[0] = CUT; bc[1] = PBC; bc[2] = PBC; };
+  virtual materialBase* lookupMaterial( double* pos );
 
   virtual void addAtomsToLayer( int layer, int n, int Z );
 };
