+++ sample_solid.cc
@@ -1,6 +1,6 @@
 #include "sample_solid.h"
 
-materialBase*  sampleSolid::lookupMaterial( float* pos )
+materialBase*  sampleSolid::lookupMaterial( double* pos )
 {
   return material[0];
 }
